Use block-by-block iteration for numeric competitive iterators
This commit introduces a SkipBlockRangeIterator, that performs better
than DocValuesRangeIterator as a competitive iterator due to more
useful docIdRunEnd() and intoBitSet() implementations.
